se acomodan Enum para recibir String y transfromarlo

+++ src/main/java/unlp/info/mapaw/comedor/domain/Ingredient.java
@@ -1,5 +1,7 @@
 package unlp.info.mapaw.comedor.domain;
 
+import java.math.BigDecimal;
+
 import javax.persistence.Entity;
 import javax.persistence.JoinColumn;
 import javax.persistence.ManyToOne;
@@ -15,13 +17,13 @@ public class Ingredient extends AbstractEntity {
 	@JoinColumn(name = "id_ingredient_recipe")
 	private IngredientRecipe recipe;
 
-	private int quantity;
+	private BigDecimal quantity;
 
-	public int getQuantity() {
+	public BigDecimal getQuantity() {
 		return quantity;
 	}
 
-	public void setQuantity(int quantity) {
+	public void setQuantity(BigDecimal quantity) {
 		this.quantity = quantity;
 	}
